requirements.txt
google-generativeai>=0.1.0rc1
gspread>=5.12.2
langid>=1.1.6
google-auth>=2.24.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
playwright>=1.40.0

